% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_impact.R
\name{load_impact}
\alias{load_impact}
\title{Load IMPACT Data}
\usage{
load_impact(
  dmp_ids,
  impact_dir = "/data1/shahs3/users/zatzmanm/work/repos/shared/msk-impact/",
  onckb_dir =
    "/data1/shahs3/users/zatzmanm/work/repos/shared/oncokb-annotated-msk-impact",
  facets_data = NULL,
  save_to = NULL
)
}
\arguments{
\item{dmp_ids}{Vector of DMP IDs (e.g. 'P-0035733-T02-IM6')}

\item{impact_dir}{IMPACT directory location}

\item{save_to}{Optional. Path to save the impact object rds (e.g. 'data/impact_data.rds')}
}
\value{
A list containing: clin_dat_patient, clin_dat_sample, onco_meta, segments, data_cna, mutations, sv, onco_dat_full
}
\description{
Load IMPACT Data
}
